🐛(oulad) remove OULAD dataset alteration from get_oulad
The get_oulad function modified the original OULAD dataset which is not
wanted.
E.g. the week_from and week_to columns were removed from the vle table
and are used in some studies that account for vle ressource
availability.
Thus we choose to no longer alter the OULAD dataset in the get_oulad
function.