⚡️(project) cache jupyter notebook cell results with CacheMagic
We want to improve the execution time of notebooks during jupyterbook
build.
Previously we configured jupyterbook to cache executed notebooks.
However, when a cell changes in a cached notebook, it's reexecuted
entirely, which is timeconsuming.
Thus we choose to implement a %%cache magic function that enables to
selectively persist variables of a notebook cell in pickle files.
We also update our notebooks in consequence to use the %%cache magic.